Switch to stateless A/B testing
This commit switches from using cookies for managing A/B testing on GOV.UK to using a stateless option based on the information provided in HTTP requests. Specifically, this approach uses the client IP address and user-agent fields from the HTTP request, taking the MD5 hash of these to use as the seed for the A/B switching logic (`randombool_seeded`).

This approach is [detailed by Fastly on their blog](https://www.fastly.com/blog/ab-testing-edge).